file structure [refactor directory and filename]
